git commit -m "Major server-side refactor and feature additions
- Restructure server configuration:
  * Add separate files for middleware, routes, database, and session setup
- Enhance TypeScript support:
  * Update tsconfig.json
  * Add custom type definition for csv-parser
- Improve error handling and logging:
  * Implement new logging system with combined.log and error.log
- Add new features:
  * Implement WebSocket support
  * Add CSV handling utilities
  * Create environment and test page routes
- Update existing components:
  * Refactor auth controller and middleware
  * Modify cart and user models
  * Enhance cart and payment services
- Implement new utility functions:
  * Add CORS, session, and validation utilities
- Update dependencies in package.json and package-lock.json

This commit significantly improves code organization, type safety, and adds
new functionality to the server-side application."

+++ server/src/services/paymentService.ts
@@ -17,7 +17,7 @@ export const processPayment = async (
 ): Promise<CreatePaymentResponse['payment']> => {
   const idempotencyKey = crypto.randomBytes(12).toString('hex');
   const paymentMoney = {
-    amount: Math.round(amount * 100), // Square API expects amount in cents
+    amount: BigInt(Math.round(amount * 100)), // Convert to BigInt, Square API expects amount in cents
     currency,
   };
